STERGM EGMoME now supports duration target statistics: * SAN failing in stergm.EGMoME() now fails gracefully. (This is necessary since not all possible target statistics can be SAN-ed. (Yet.)) * InitErgmTerm.duration.R has been added, containing the new duration target statistics:  * edge.ages = sum over extant edges of the times since they were formed,  * edgecov.ages = weighted sum over the same, and  * edges.ageinterval which can now be used as a target statistic as well, or a dissolution statistic. * ergm.getmodel() now takes an additional optional role= argument, with values being "static", "formation", "dissolution", and "target", to give change statistics that want it a hint of how they are being used and to prevent their being used in an unsupported role. Some new statistics (like edge.ages) can only be targets, while others (edges.ageinterval) use different C functions depending on whether they are used as targets or generatively. The change is transparent to ordinary ERGM statistics and calls. * On the C side, dynamic MCMC code now supports timer (t_???) statistics, which are called after the time step's toggles have been made but before the clock advances. These allow the code to keep track of above-mentioned duration-dependent statistics, allowing them to be used as targets. * ergm.getglobalstats can (via ergm.Cprepare, where it was implemented a while ago) support "time" and "lasttoggle" attributes (paralleling the structures in edgetree.h) attached to an ordinary network. These are for internal use only, and may eventually be replaced by networkDynamic's facilities.
simulate.network (STERGM):
* nsim > 1 now produces multiple independent network simulations.
* An argument time.points specifies the number of time points (successive, if time.interval=1) whose information is to be returned.
* toggles= argument has been replaced with statsonly=, which works the same way as in simulate.formula() (ERGM). It defaults to TRUE if time.burnin>0 or time.interval>1.
* There are two new arguments, stats.form= and stats.diss=, both defaulting to FALSE, controlling whether formation and dissolution statistics will be tracked and returned. Setting them to TRUE is not recommended, especially if you are using duration-dependent statistics.
* Argument monitor= now accepts a string, just like targets= argument of stergm(): "formation" to copy from formation, "dissolution" from dissolution, and "all" to combine the unique statistics of the two. This is the preferred way to collect statistics of interest.
* control.simulate.stergm/.network() now takes the proper MCMC.prop.* arguments.
* If statsonly=TRUE, the function returns a list with at most three elements: stats.form, stats.diss, and stats.mon.
* If statsonly=FALSE, it returns a networkDynamic with (attr()-style) attributes analagous to those of simulate.formula() (ERGM) attached describing the attributes of the simulation. There is no print method that makes use of these attributes at this time.
* dynamic_simtest.R test has been updated, but doesn't test for as much as it should.

Other fixes and cleanups:
* InitErgm.duration.R has been removed.	
* Statistics D.off, D.dyad, and D.on lacked apparent purpose and have been removed. (We can always put them back if needed.)


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@11426 9cb281f4-553f-406f-8d7f-11f072bbfe46

+++ src/changestats.c
@@ -1834,7 +1834,7 @@ D_CHANGESTAT_FN(d_dyadcov) {
 D_CHANGESTAT_FN(d_edgecov) {
   double val;
   Vertex tail, head;
-  long int nrow, noffset;
+  int nrow, noffset;
   int i, edgeflag;
   
   noffset = BIPARTITE;
